feat(snackbar): introduce MSnackbarProvider and state management
Refactor Snackbar to use a dedicated provider for state management and
simplify MSnackbarHost. Renamed MSnackbarHostState to MSnackbarState
and MSnackbar to MSnackbarContent for clarity.
